# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DemoOrderIn": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,18}|(?=[\\d.]{1,19}0*$)\\d*\\.\\d*0*$)",
                "type": "string"
              }
            ],
            "default": "20000",
            "title": "Amount"
          },
          "currency": {
            "default": "INR",
            "pattern": "^[A-Za-z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "customer_id": {
            "default": "C-17",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,63}$",
            "title": "Customer Id",
            "type": "string"
          }
        },
        "title": "DemoOrderIn",
        "type": "object"
      },
      "FaultIn": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "maxLength": 64,
            "title": "Kind",
            "type": "string"
          },
          "order_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "times": {
            "default": 1,
            "maximum": 100.0,
            "minimum": -1.0,
            "title": "Times",
            "type": "integer"
          }
        },
        "required": [
          "kind"
        ],
        "title": "FaultIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Payment execution, recovery and reconciliation between applications or AI agents and a payment provider. Contract: docs/API.md. Simulated provider only.",
    "title": "IntentGuard Gateway",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/dev/faults": {
      "delete": {
        "operationId": "clear_faults_api_dev_faults_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Clear Faults Api Dev Faults Delete"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Clear provider faults",
        "tags": [
          "dev (simulator only)"
        ]
      },
      "get": {
        "operationId": "list_faults_api_dev_faults_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Faults Api Dev Faults Get"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Queued provider faults",
        "tags": [
          "dev (simulator only)"
        ]
      },
      "post": {
        "operationId": "inject_fault_api_dev_faults_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FaultIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Inject Fault Api Dev Faults Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inject a provider fault",
        "tags": [
          "dev (simulator only)"
        ]
      }
    },
    "/api/dev/ledger": {
      "get": {
        "operationId": "ledger_api_dev_ledger_get",
        "parameters": [
          {
            "in": "query",
            "name": "order_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Ledger Api Dev Ledger Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provider ground truth (every transaction, ignoring search visibility)",
        "tags": [
          "dev (simulator only)"
        ]
      }
    },
    "/api/dev/orders": {
      "post": {
        "operationId": "demo_order_api_dev_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoOrderIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Demo Order Api Dev Orders Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a fresh demo order (so demo scenarios can be repeated)",
        "tags": [
          "dev (simulator only)"
        ]
      }
    },
...
  }
}